requirements.txt
django
django-constance
django-markdownify
django-picklefield
reportlab
